Change all file handlings to streams
Everything related to files is now handled by streams.
This gives huge memory savings and unlimited script sizes (just limited by object counts in scripts).

+++ src/firmware/src/core/message/messages/mesh_connection.hpp
@@ -16,8 +16,8 @@ namespace SyncBlink
         bool connectedToMeshWifi;
         uint64_t parentId;
         uint16_t ledCount;
-        uint8_t majorVersion;
-        uint8_t minorVersion;
+        uint16_t majorVersion;
+        uint16_t minorVersion;
         std::string nodeLabel;
     };
